test spark with mlflow
